Update patch set 1
Patch Set 1:

I tested the ISO build on both x86_64 and aarch64 and it builds successfully.
However, I haven't yet tested the deployment - if you think that is necessary (not sure whether other changes are required in TA/REC, although I expect not to), I'll get back with the result of that test too.

Patch-set: 1